week_2_whisperTask/requirements.txt
openai-whisper==20231117
deep-translator==1.11.4
ffmpeg-python==0.2.0
